Making current global / tracking integer (mA instead of A). Something else small.

+++ sensors/sensor_api.h
@@ -46,7 +46,7 @@ float getCurrentTotal();
 uint8_t getAcceleration();
 
 // Update threshold values that trigger an eStop
-void setThresholdTemp(uint16_t threshTemp);
+void setThresholdTemp(uint8_t threshTemp);
 void setThresholdCurrent(uint16_t threshCurrent);
 void setThresholdAccel(uint16_t threshAccel);
